+++ spec/support/sandbox_project.rb
@@ -13,6 +13,7 @@ require "json"
 class SandboxProject
   LIB = File.expand_path("../../lib", __dir__)
   EXECUTABLE = File.expand_path("../../exe/rspec-signal", __dir__)
+  PARALLEL_EXECUTABLE = File.expand_path("../../exe/rspec-signal-parallel", __dir__)
 
   attr_reader :root
 
@@ -54,6 +55,12 @@ class SandboxProject
     stdout, stderr, status = Open3.capture3(
       env, RbConfig.ruby, EXECUTABLE, "--require", "./spec/spec_helper.rb", "--no-color", *args, chdir: root
     )
+    Run.new(stdout: stdout, stderr: stderr, status: status.exitstatus, project: self)
+  end
+
+  def run_signal_parallel(*args)
+    env = { "RUBYLIB" => [LIB, ENV.fetch("RUBYLIB", nil)].compact.join(File::PATH_SEPARATOR) }
+    stdout, stderr, status = Open3.capture3(env, RbConfig.ruby, PARALLEL_EXECUTABLE, *args, chdir: root)
     Run.new(stdout: stdout, stderr: stderr, status: status.exitstatus, project: self)
   end